+++ laktory/models/compute/__init__.py
@@ -1,0 +1,3 @@
+from laktory.models.compute.initscript import InitScript
+from laktory.models.compute.cluster import Cluster
+from laktory.models.compute.warehouse import Warehouse