+++ laktory/models/secrets/__init__.py
@@ -1,0 +1,2 @@
+from laktory.models.secrets.secret import Secret
+from laktory.models.secrets.secretscope import SecretScope