+++ laktory/models/secrets/secret.py
@@ -1,0 +1,7 @@
+from laktory.models.base import BaseModel
+
+
+class Secret(BaseModel):
+    scope: str = None
+    key: str = None
+    value: str = None