+++ laktory/resourcesengines/pulumi/__init__.py
@@ -3,5 +3,6 @@ from laktory.resourcesengines.pulumi.group import PulumiGroup
 from laktory.resourcesengines.pulumi.schema import PulumiSchema
 from laktory.resourcesengines.pulumi.serviceprincipal import PulumiServicePrincipal
 from laktory.resourcesengines.pulumi.user import PulumiUser
+from laktory.resourcesengines.pulumi.cluster import PulumiCluster
 from laktory.resourcesengines.pulumi.volume import PulumiVolume
 from laktory.resourcesengines.pulumi.base import PulumiResourcesEngine